# post-data-management-back-end-development/docs/swagger.yaml
basePath: /v1
definitions:
  apierrors.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierrors.AppError'
      message:
        description: Descriptive message providing details about the error.
        type: string
      status_code:
        description: HTTP status code representing the error type (e.g., 404 for Not
          Found).
        type: integer
      success:
        description: Success flag to indicate if the operation was successful.
        type: boolean
    type: object
  apierrors.AppError:
    properties:
      code:
        type: string
      field_errors:
        items:
          $ref: '#/definitions/apierrors.FieldError'
        type: array
      id:
        type: string
      message:
        type: string
    type: object
  apierrors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  domain.CadreMasterNew:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
    type: object
...
  domain.DesignationMasterNew:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      designation:
        type: string
      designation_id:
        type: integer
    type: object
...
  handler.CreateDesignationRequest:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      created_by:
        type: string
      designation:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      remarks:
        type: string
      status:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - cadre_id
    - cadre_name
    - created_by
    - designation
    - group_id
    - group_name
    - remarks
    - status
    - valid_from
    - valid_to
    type: object
  handler.DesignationMasterDetails:
    properties:
      designation_id:
        type: integer
      designation_name:
        type: string
    type: object
...
  handler.UpdateDesignationRequest:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      designation:
        type: string
      designation_id:
        type: integer
      designationUID:
        type: integer
      group_id:
        type: integer
      group_name:
        type: string
      remarks:
        type: string
      status:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - cadre_id
    - cadre_name
    - designation
    - designationUID
    - designation_id
    - group_id
    - group_name
    - remarks
    - status
    - valid_from
    - valid_to
    type: object
...
  port.MetaDataResponse:
    properties:
      limit:
        type: integer
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      total_records_count:
        type: integer
    type: object
  port.StatusCodeAndMessage:
    properties:
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  response.CadreListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.CadreMasterNew'
        type: array
      metadata:
        $ref: '#/definitions/port.MetaDataResponse'
      status:
        $ref: '#/definitions/port.StatusCodeAndMessage'
    type: object
...
  response.CreateDesignationMasterAPIResponse:
    properties:
      data:
        $ref: '#/definitions/response.CreateDesignationMasterResponse'
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  response.CreateDesignationMasterResponse:
    properties:
      designation_id:
        type: integer
      designation_uid:
        type: integer
    type: object
...
  response.DesignationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.DesignationMasterNew'
        type: array
      metadata:
        $ref: '#/definitions/port.MetaDataResponse'
      status:
        $ref: '#/definitions/port.StatusCodeAndMessage'
    type: object
...
  response.ListAllDesignationsAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ListAllDesignationsResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
...
  response.ListAndFilterDesignationsAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ListAndFilterDesignationsResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.ListAndFilterDesignationsResponse:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      designation:
        type: string
      designation_id:
        type: integer
      designation_name:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support_cept@indiapost.gov.in
    name: API Support Team
    url: http://cept.gov.in/support
  description: A comprehensive API for managing Post Management Data Management, offering
    endpoints for creation, update, deletion, and retrieval of Posts Data.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://cept.gov.in/terms
  title: Post Management Data Management API
  version: "1.0"
paths:
  /designation/list:
    get:
      consumes:
      - application/json
      description: List all designations from the database
      produces:
      - application/json
      responses:
        "200":
          description: List all designations from the database
          schema:
            items:
              $ref: '#/definitions/handler.DesignationMasterDetails'
            type: array
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: ListAllDesignations
      tags:
      - designationMaster
  /designations/cadres-by-group:
    get:
      consumes:
      - application/json
      description: Fetch a list of cadres belonging to a specific group, with optional
        pagination.
      parameters:
      - description: Group Code (e.g., A, B, C, D)
        in: query
        name: group_code
        required: true
        type: string
      - description: Number of records to skip (for pagination)
        in: query
        name: skip
        type: integer
      - description: Number of records to return (for pagination)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cadres fetched successfully
          schema:
            $ref: '#/definitions/response.CadreListResponse'
        "400":
          description: Validation or binding error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Cadres by Group
      tags:
      - Designation Master
  /designations/designations-by-cadre:
    get:
      consumes:
      - application/json
      description: Fetch a list of designations belonging to a specific cadre, with
        optional pagination.
      parameters:
      - description: Cadre Code (e.g., ENG, CLERK)
        in: query
        name: cadre-code
        required: true
        type: string
      - description: Number of records to skip (for pagination)
        in: query
        name: skip
        type: integer
      - description: Number of records to return (for pagination)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Designations fetched successfully
          schema:
            $ref: '#/definitions/response.DesignationListResponse'
        "400":
          description: Validation or binding error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Designations by Cadre
      tags:
      - Designation Master
...
  /post-management/designations:
    get:
      consumes:
      - application/json
      description: Fetches the list of designations filtered by group and cadre codes.
        Also returns the complete list of designation types.
      parameters:
      - description: Group Code (optional)
        in: query
        name: group-code
        type: string
      - description: Cadre Code (optional)
        in: query
        name: cadre-code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Filtered designations and all designation types retrieved successfully
          schema:
            $ref: '#/definitions/response.ListAndFilterDesignationsAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get and Filter Designations
      tags:
      - Designation Management
  /post-management/designations/{designation-uid}:
    put:
      consumes:
      - application/json
      description: Updates a post designation master entry.
      parameters:
      - description: Designation UID to update
        in: path
        name: designation-uid
        required: true
        type: integer
      - description: designation details to update
        in: body
        name: UpdateDesignationRequest
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDesignationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: resource updated successfully
          schema:
            $ref: '#/definitions/response.CreateDesignationMasterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Conflict in the data
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Update Designation Master
      tags:
      - Designation Management
  /post-management/designations/all-designations:
    get:
      consumes:
      - application/json
      description: List all designation types or filter by group-code if provided
      parameters:
      - description: Filter designations by group code
        in: query
        name: group-code
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sort_type
        type: string
      - in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Designation list retrieved successfully
          schema:
            $ref: '#/definitions/response.ListAllDesignationsAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: List All Designation Details
      tags:
      - Designation Management
  /post-management/designations/create:
    post:
      consumes:
      - application/json
      description: Creates a post designation master entry.
      parameters:
      - description: designation details to create
        in: body
        name: CreateDesignationRequest
        required: true
        schema:
          $ref: '#/definitions/handler.CreateDesignationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Designation master details created successfully
          schema:
            $ref: '#/definitions/response.CreateDesignationMasterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Conflict in the data
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Create Designation Master
      tags:
      - Designation Management
...
schemes:
- http
- https
swagger: "2.0"